Throw an error if no stream size information
Currently, this module attempts to get size information from a `length`
property, or a `content-length` HTTP header if coming from Resin
Request.

The CLI attempts to always display a progress bar when writing an image,
therefore the absence of this property represents a UX error (a spinner
is not enough in this case, since the whole purpose of building this
module is to be able to provide progress of the operation).

Resin Request is being updated to v2.2.5 in this PR in order to get bug
fixes needed for this change to work.