Make write check work in OS X and Windows
Currently the `.check()` function failed on OS X and Windows for images
that contain FAT partitions, since these operating systems will
auto-mount the FAT partitions when the drive is complete, and "touch"
some files like `.DS_Store`, causing the checksum to be slightly
different.

This PR also removes `.check()` and instead adds a `check` option to
`.write()`. This is because we need more control on the time slice after
the writing is complete and the check process is started.

OS X
----

For OS X, we use a module called denymount
(https://github.com/resin-io/denymount), which causes the OS to omit
auto-mounting the drives we specify.

Windows
-------

In Windows, we triggered `diskpart rescan` when the writing finished in
order for the OS to catch up with the changes. Turns out we can delay
this until *after* the check, so the new FAT partitions are not
mounted.