Allow to pass a transform stream to the writer
This transform stream is applied after the progress stream, but before
the actual writing.

Its very useful to handle compressed files where you don't know the
uncompressed size in advance.

For example, we can pass an xz compressed image as a source stream to
`.write()`, the compressed size as `options.size`, and a LZMA
decompressor transform stream as `options.transform`.

If this option is not passed, the transform stream defaults to
`stream.PassThrough`, which basically does nothing.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>