Fix Windows EPERM on GPT drives
This commit fixes a sporadic issues observed by a minority of our
Windows users triggered when writing beyond the first 512 bytes of a
drive.

After some experimentation, the issue comes down to writing an image to
a drive previously formatted with a GUID Partition Table. This
technology stores its partition data in various blocks at both the
beginning and end of a drive. MBR, in the other hand, only stores the
partition table in the first block. We were currently only accounting
for MBR drives.

Windows only allows direct write to physical drives if the section we're
writing to is the first block of the drive, or if the drive contains no
file system (a "RAW" drive, as Microsoft describes it).

Since deleting the first block of a GPT drive will not delete all the
partition table information, Windows didn't detect it as "RAW", and
therefore threw `EPERM` when we wrote beyond the first block.

Sadly, doing manual I/O to clean the last blocks of drive was not
possible, since Windows would only allow us to write to the first block.
We instead make use of `diskpart`'s `clean` command, which bypasses this
Windows limitation by requesting an exclusive lock on the drive
(something we can't do without interacting with the Windows API
 directly), and knows how to clean both MBR and GPT drives.

See: https://github.com/resin-io/etcher/issues/334
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>