Don't throw ENOSPC if the drive capacity is equal to the image size
The code that checks whether we're about to run out of space has an off
by one issue.

The function collects the length of the chunks the transform stream
received as `transferredBytes`. Once the variable has been incremented,
it checks for

```
transferredBytes + options.chunkSize > options.driveSize
```

But it makes no sense to add the chunk size to `transferredBytes` since
we just incremented it before, causing the variable to be incremented
twice, and give false alarms when the image size is very close to the
capacity of the drive.

See: https://github.com/resin-io/etcher/issues/629
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>