Allow a minimum chunk size rather than enforcing a big one
Consider an image which size is 2621440 bytes, which equals to 2.5 MB.
This module enforces a block size of 1 MB, meaning that when processing
the last chunk (512 KB), it will get padded by another 512 KB.

This introduces the chance that the drive might not be able to hold the
padded image, but has enough space to hold to unpadded image, causing a
lot of confusion to the user.

In order to mitigate this problem, but still preserve the default big
block size (for speed reasons), we introduce the concept of a "minimum
block size". What this means is that the input stream is automatically
chunked to the default block size (e.g: 1 MB), but if a chunk is smaller
than the block size, its left alone *unless* its smaller than the
minimum block size, in which case its padded to match the minimum block
size instead.

Basically, this means that 99% of the blocks will be padded to 1 MB, but
some of them, in rare ocasions where the image is not 1MB-aligned and
the drive is almost the same size as the image, they will be padded to
512 KB.

If a chunk size is bigger than the minimum block size, there is still a
chance that its very malformed, and could cause issues when being copied
to the actual drive. As workaround, blocks that are bigger that our
minimum block size get padded to match the initial block size.

See: https://github.com/resin-io/etcher/issues/685
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>