Validate contents using the actual transferred bytes information
Currently, we slice the flashed drive stream to the image size number in
order to not check more data than we should, however in the case of
compressed images, the passed image size might refer to the "compressed"
image size, causing this module to slice less than it should, and
therefore cause incorrect validation errors.

As a real fix, we keep an eye on how many bytes we actually transferred,
and use that information to correctly slice the drive stream.

See: https://github.com/resin-io/etcher/issues/710
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>