Fix sporadic EPERM write errors on Windows
We've been experiencing sporadic EPERM issues on certain writes. The bug
is not easy to reproduce, and seems to randomly happen on any position
of the image. As a way to test this, we spawn several instances of the
`example.js` script shipped in this project, which increases the odd of
seeing the bug.

To offer some background, Windows only allows clients to write to
physical drives directly in the following conditions:

- The sectors to be written to are boot sectors.
- The sectors to be written to reside outside of file system space.
- You have explicitly locked or dismounted the volume by using
  `FSCTL_LOCK_VOLUME` or `FSCTL_DISMOUNT_VOLUME`.
- The volume has no actual file system.

Before this module starts writing data to the physical drive, it cleans
the drive using `diskpart.exe`, which causes the drive to lose any file
system information, and therefore allowing us to write to it (since
Windows permits writes to volumes that have no actual file system).

Given that we write in a linear fashion, the first chunk that we write
represents the partition table. In some cases, after a while, Windows
seems to realise there is a file system on the drive we're currently
writing to (since we initially wrote a partition table), and suddenly
throws EPERM, since we're violating its policies.

Once we get the EPERM, Windows doesn't allow us to write to *any* part
of the drive, except boot sectors, and re-trying has no effect.

As a way to workaround this, we take the following approach during the
writing phase: we omit the first chunk, which contains the partition
table, but temporarily save it in memory. We then proceed to write all
but the first chunk, and once all writes completes, we finally write the
first block.

This ensures that Windows will not detect any valid partition table
until the end of the write, preventing it from revoking us write access
to the volume.

Given that we clean the volume with `diskpart.exe` before starting the
write, we don't need to concern with the existent data on the first
sector when omitting the first chunk.

A more elegant approach could be making use of the `FSCTL_LOCK_VOLUME`
or `FSCTL_DISMOUNT_VOLUME` calls, however they only seem to be available
when using the Windows C API. There might be a way to simulate the
behaviour in plain JavaScript, or by implementing a NodeJS Addon, which
are approaches worth investigating in depth.

See: https://github.com/resin-io/etcher/issues/627
See: https://msdn.microsoft.com/en-us/library/aa365747(VS.85).aspx
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>