Don't consider buffer alignment bytes during the validation phase
Consider an image that is exactly 2.5M. Since the default block size
of this module is 1M, it means that the actual data that is going to be
written to the device is 3M (2.5M + 0.5M padding).

Since we collect the amount of bytes we write as `transferredBytes` in
the write routine, and then pass that exact number to the validation
phase, it means that the validation phase will consider the padding when
calculating the checksum, and therefore report incorrect validation
errors.

As a solution, we only collect the real transferred bytes in
`transferredBytes`, and we introduce a new `paddingBytes` accumulator to
store padding bytes, if any.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>