Make utils.safePipe adhere to NodeJS callback convention
Currently, `utils.safePipe()` accepts a callback that is only called
when an error happens, leaving the responsibility to attach an event
listener to the last stream to the user.

Now, `utils.safePipe()` attaches prefedefined event listeners to the
last stream (`finish`, and `done`), and calls the callback when they are
emitted.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>