Prevent unhandled error events from the first stream of `utils.safePipe`
We have a utility function called `utils.safePipe()` that accepts an
array of streams and pipes them all together attaching error handlers at
every step.

We didn't recognise a subtle flaw that would prevent the error handler
(and any event listener, really) to be attached to the first stream on
the pipe chain: we check if `current` exists, and if not, set the
current stream as `current` however we `return` right away, which means
the error handler and event handlers defined just below never get called
on the first stream (which is the only case where `current` is not
defined).

As a bonus, this commit adds some unit tests for `utils.safePipe()`.

See: resin-io/etcher#981
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>