Correctly throw TypedError instances
This module declares `error.ValidationError`, which is an instance of
`TypedError`. We used it like this:

```js
throw errors.ValidationError;
```

The resulting thrown error is not an error instance, but still has some
of the properties we declare, like the code, therefore confusing the
tests.

This eventually caused problems in Etcher when trying to stringify the
error using the robot module.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>